# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Outlier AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/meta/accounts": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List Accounts",
        "operationId": "list_accounts_api_meta_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountOut"
                  },
                  "type": "array",
                  "title": "Response List Accounts Api Meta Accounts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Create Account",
        "operationId": "create_account_api_meta_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/accounts/{account_id}": {
      "put": {
        "tags": [
          "meta"
        ],
        "summary": "Update Account",
        "operationId": "update_account_api_meta_accounts__account_id__put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/sync": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Sync Insights",
        "operationId": "sync_insights_api_meta_sync_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync Insights Api Meta Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/sync_comments": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Sync Comments",
        "operationId": "sync_comments_api_meta_sync_comments_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync Comments Api Meta Sync Comments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/kill_switch": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Kill Switch",
        "operationId": "get_kill_switch_api_meta_kill_switch_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "meta"
        ],
        "summary": "Set Kill Switch",
        "operationId": "set_kill_switch_api_meta_kill_switch_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "meta_account_id": {
            "type": "string",
            "title": "Meta Account Id"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          },
          "pixel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pixel Id"
          },
          "attribution_setting": {
            "type": "string",
            "title": "Attribution Setting",
            "default": "7d_click_1d_view"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "daily_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Cap Usd"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Token"
          },
          "is_fake": {
            "type": "boolean",
            "title": "Is Fake",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "meta_account_id"
        ],
        "title": "AccountCreate"
      },
      "AccountOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "meta_account_id": {
            "type": "string",
            "title": "Meta Account Id"
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id"
          },
          "pixel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pixel Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attribution_setting": {
            "type": "string",
            "title": "Attribution Setting"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "daily_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Cap Usd"
          },
          "is_fake": {
            "type": "boolean",
            "title": "Is Fake"
          },
          "has_access_token": {
            "type": "boolean",
            "title": "Has Access Token",
            "default": false
          },
          "has_page_token": {
            "type": "boolean",
            "title": "Has Page Token",
            "default": false
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "meta_account_id",
          "page_id",
          "pixel_id",
          "status",
          "attribution_setting",
          "api_version",
          "category",
          "daily_cap_usd",
          "is_fake",
          "last_sync_at",
          "last_error"
        ],
        "title": "AccountOut"
      },
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(active|needs_reauth|disabled)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "daily_cap_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Cap Usd"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Token"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "title": "AccountUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KillSwitchIn": {
        "properties": {
          "shipping_enabled": {
            "type": "boolean",
            "title": "Shipping Enabled"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "shipping_enabled"
        ],
        "title": "KillSwitchIn"
      },
      "KillSwitchOut": {
        "properties": {
          "shipping_enabled": {
            "type": "boolean",
            "title": "Shipping Enabled"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "changed_by": {
            "type": "string",
            "title": "Changed By"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "shipping_enabled",
          "reason",
          "changed_by",
          "changed_at"
        ],
        "title": "KillSwitchOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}